Added instructions/settings tabs, generic location validation
Added buttons that open up tabs for instructions and for settings
(actual instructions and settings not present yet). Added validation to
generic location to verify user input matches one of the API options for
generic locations. Users are now also allowed to enter locations such as
"Hardware Store" which will be automatically converted to the API
readable "hardware_store"